python-backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==1.17.1
click==8.2.1
dotenv==0.9.9
fastapi==0.116.1
functions==0.7.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
packaging==25.0
pluggy==1.6.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
python-multipart==0.0.20
schemas==0.7.1
sniffio==1.3.1
split==0.4
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
